feat: 스토리 도메인 작성 및 S3 연동

+++ src/main/java/io/itmca/lifepuzzle/domain/story/endpoint/response/StoryTagResponse.java
@@ -9,10 +9,10 @@ import lombok.*;
 public class StoryTagResponse implements Comparable<StoryTagResponse> {
     String key;
     String displayName;
-    int priority;
+    Long priority;
 
     @Override
     public int compareTo(StoryTagResponse o) {
-        return Integer.compare(this.priority, o.priority);
+        return Long.compare(this.priority, o.priority);
     }
 }
